Simple Test Functionality
Using test db and data, the main page will populate a div
dynamically with the results of a query, only taking in
the time selection. currently, only times 2 and 3 show
anything. Note on the code: the result array is indexed by
column, but does not track table it came from. this could
cause errors if tables have identical column names. Casting
a result with an alias may remedy that in the php, but I
have not tested that.